+++ common-lootjs/src/main/java/com/yanny/ali/plugin/lootjs/node/ItemTagNode.java
@@ -60,10 +60,10 @@ public class ItemTagNode implements IDataNode, IItemNode {
         tooltip = NodeUtils.decodeTooltipNodes(utils, buf);
         count = new RangeValue(buf);
         modified = buf.readBoolean();
+        chance = buf.readFloat();
 
         conditions = Collections.emptyList();
         functions = Collections.emptyList();
-        chance = 1;
     }
 
     public boolean isModified() {
@@ -101,6 +101,7 @@ public class ItemTagNode implements IDataNode, IItemNode {
         NodeUtils.encodeTooltipNodes(utils, buf, tooltip);
         count.encode(buf);
         buf.writeBoolean(modified);
+        buf.writeFloat(chance);
     }
 
     @Override