+++ common/src/main/java/com/yanny/ali/manager/AliServerRegistry.java
@@ -64,6 +64,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
     private final Set<Class<?>> missingConditionTooltips = new HashSet<>();
     private final Set<Class<?>> missingIngredientTooltips = new HashSet<>();
     private final Set<Class<?>> missingItemListingFactories = new HashSet<>();
+    private final Set<Class<?>> missingNumberConverters = new HashSet<>();
 
     private final ICommonUtils utils;
 
@@ -100,6 +101,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         missingConditionTooltips.clear();
         missingIngredientTooltips.clear();
         missingItemListingFactories.clear();
+        missingNumberConverters.clear();
     }
 
     public void addLootTable(ResourceLocation resourceLocation, LootTable lootTable) {
@@ -371,7 +373,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
                     LOGGER.warn("Failed to convert number with error {}", throwable.getMessage());
                 }
             } else {
-                LOGGER.warn("Number converter for {} was not registered", numberProvider.getClass().getCanonicalName());
+                missingNumberConverters.add(numberProvider.getClass());
             }
         }
 
@@ -435,6 +437,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         missingConditionTooltips.forEach((t) -> LOGGER.warn("Missing condition tooltip for {}", t.getName()));
         missingIngredientTooltips.forEach((t) -> LOGGER.warn("Missing ingredient tooltip for {}", t.getName()));
         missingItemListingFactories.forEach((t) -> LOGGER.warn("Missing trade item listing for {}", t.getName()));
-        LOGGER.info("Prepared {} loot tables", lootTableMap.size());
+        missingNumberConverters.forEach((t) -> LOGGER.warn("Missing number converters for {}", t.getName()));
+        LOGGER.info("Prepared {} loot tables and {} trades", lootTableMap.size(), itemListingFactoryMap.size());
     }
 }