+++ common-jei/src/main/java/com/yanny/ali/compatibility/JeiCompatibility.java
@@ -5,10 +5,10 @@ import com.yanny.ali.Utils;
 import com.yanny.ali.api.IDataNode;
 import com.yanny.ali.compatibility.common.*;
 import com.yanny.ali.compatibility.jei.*;
+import com.yanny.ali.configuration.LootCategory;
 import com.yanny.ali.manager.AliClientRegistry;
+import com.yanny.ali.manager.AliCommonRegistry;
 import com.yanny.ali.manager.PluginManager;
-import com.yanny.ali.registries.LootCategories;
-import com.yanny.ali.registries.LootCategory;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -23,30 +23,25 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.block.Block;
-import org.apache.commons.lang3.mutable.MutableObject;
 import org.apache.commons.lang3.tuple.Pair;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
+import java.util.stream.Collector;
 import java.util.stream.Collectors;
 
 @JeiPlugin
 public class JeiCompatibility implements IModPlugin {
     private static final Logger LOGGER = LogUtils.getLogger();
 
-    private final Map<LootCategory<Block>, JeiBlockLoot> blockCategories = new HashMap<>();
-    private final Map<LootCategory<EntityType<?>>, JeiEntityLoot> entityCategories = new HashMap<>();
-    private final Map<LootCategory<String>, JeiGameplayLoot> gameplayCategories = new HashMap<>();
-    private final Map<LootCategory<String>, JeiTradeLoot> tradeCategories = new HashMap<>();
-
-    private final MutableObject<JeiBlockLoot> defaultBlockLoot = new MutableObject<>();
-    private final MutableObject<JeiEntityLoot> defaultEntityLoot = new MutableObject<>();
-    private final MutableObject<JeiGameplayLoot> defaultGameplayLoot = new MutableObject<>();
-    private final MutableObject<JeiTradeLoot> defaultTradeLoot = new MutableObject<>();
+    private final Map<LootCategory<Block>, JeiBlockLoot> blockCategories = new LinkedHashMap<>();
+    private final Map<LootCategory<EntityType<?>>, JeiEntityLoot> entityCategories = new LinkedHashMap<>();
+    private final Map<LootCategory<ResourceLocation>, JeiGameplayLoot> gameplayCategories = new LinkedHashMap<>();
+    private final Map<LootCategory<ResourceLocation>, JeiTradeLoot> tradeCategories = new LinkedHashMap<>();
 
     @Override
     public void onRuntimeUnavailable() {
@@ -54,15 +49,11 @@ public class JeiCompatibility implements IModPlugin {
         entityCategories.clear();
         gameplayCategories.clear();
         tradeCategories.clear();
-
-        defaultBlockLoot.setValue(null);
-        defaultEntityLoot.setValue(null);
-        defaultGameplayLoot.setValue(null);
-        defaultTradeLoot.setValue(null);
     }
 
     @Override
     public void registerCategories(IRecipeCategoryRegistration registration) {
+        AliCommonRegistry commonRegistry = PluginManager.COMMON_REGISTRY;
         IGuiHelper guiHelper = registration.getJeiHelpers().getGuiHelper();
 
         blockCategories.clear();
@@ -70,51 +61,45 @@ public class JeiCompatibility implements IModPlugin {
         gameplayCategories.clear();
         tradeCategories.clear();
 
-        blockCategories.putAll(LootCategories.BLOCK_LOOT_CATEGORIES.entrySet().stream()
-                .map((e) -> new AbstractMap.SimpleEntry<>(e.getValue(), createCategory(guiHelper, e, JeiBlockLoot::new)))
-                .collect(Collectors.toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue)));
-        entityCategories.putAll(LootCategories.ENTITY_LOOT_CATEGORIES.entrySet().stream()
-                .map((e) -> new AbstractMap.SimpleEntry<>(e.getValue(), createCategory(guiHelper, e, JeiEntityLoot::new)))
-                .collect(Collectors.toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue)));
-        gameplayCategories.putAll(LootCategories.GAMEPLAY_LOOT_CATEGORIES.entrySet().stream()
-                .map((e) -> new AbstractMap.SimpleEntry<>(e.getValue(), createCategory(guiHelper, e, JeiGameplayLoot::new)))
-                .collect(Collectors.toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue)));
-        tradeCategories.putAll(LootCategories.TRADE_LOOT_CATEGORIES.entrySet().stream()
-                .map((e) -> new AbstractMap.SimpleEntry<>(e.getValue(), createCategory(guiHelper, e, JeiTradeLoot::new)))
-                .collect(Collectors.toMap(AbstractMap.SimpleEntry::getKey, AbstractMap.SimpleEntry::getValue)));
+        blockCategories.putAll(commonRegistry.getConfiguration().blockCategories.stream().collect(getCollector(guiHelper, JeiBlockLoot::new)));
+        entityCategories.putAll(commonRegistry.getConfiguration().entityCategories.stream().collect(getCollector(guiHelper, JeiEntityLoot::new)));
+        gameplayCategories.putAll(commonRegistry.getConfiguration().gameplayCategories.stream().collect(getCollector(guiHelper, JeiGameplayLoot::new)));
+        tradeCategories.putAll(commonRegistry.getConfiguration().tradeCategories.stream().collect(getCollector(guiHelper, JeiTradeLoot::new)));
 
         blockCategories.values().forEach(registration::addRecipeCategories);
         entityCategories.values().forEach(registration::addRecipeCategories);
         gameplayCategories.values().forEach(registration::addRecipeCategories);
         tradeCategories.values().forEach(registration::addRecipeCategories);
-
-        defaultBlockLoot.setValue(getDefaultCategory(guiHelper, JeiBlockLoot::new, registration, blockCategories, LootCategories.BLOCK_LOOT));
-        defaultEntityLoot.setValue(getDefaultCategory(guiHelper, JeiEntityLoot::new, registration, entityCategories, LootCategories.ENTITY_LOOT));
-        defaultGameplayLoot.setValue(getDefaultCategory(guiHelper, JeiGameplayLoot::new, registration, gameplayCategories, LootCategories.GAMEPLAY_LOOT));
-        defaultTradeLoot.setValue(getDefaultCategory(guiHelper, JeiTradeLoot::new, registration, tradeCategories, LootCategories.TRADE_LOOT));
     }
 
     @Override
     public void registerRecipes(IRecipeRegistration registration) {
-        CompletableFuture<Pair<Map<ResourceLocation, IDataNode>, Map<ResourceLocation, IDataNode>>> futureData = new CompletableFuture<>();
+        CompletableFuture<Pair<Map<ResourceLocation, IDataNode>, Map<ResourceLocation, IDataNode>>> futureData = PluginManager.CLIENT_REGISTRY.getCurrentDataFuture();
 
-        PluginManager.CLIENT_REGISTRY.setOnDoneListener((lootData, tradeData) -> futureData.complete(Pair.of(lootData, tradeData)));
-
-        if (!futureData.isDone()) {
+        if (futureData.isDone()) {
+            LOGGER.info("Data already received, processing instantly.");
+        } else {
             LOGGER.info("Blocking this thread until all data are received!");
         }
 
         try {
-            Pair<Map<ResourceLocation, IDataNode>, Map<ResourceLocation, IDataNode>> pair = futureData.get(30, TimeUnit.SECONDS);
+            Pair<Map<ResourceLocation, IDataNode>, Map<ResourceLocation, IDataNode>> pair = futureData.get();
 
             registerData(registration, pair.getLeft(), pair.getRight());
-        } catch (TimeoutException e) {
-            futureData.cancel(true);
-            PluginManager.CLIENT_REGISTRY.clearLootData();
-            LOGGER.error("Failed to received data in 30 seconds, registration aborted!");
+        } catch (ExecutionException e) {
+            Throwable cause = e.getCause() != null ? e.getCause() : e;
+
+            if (cause instanceof TimeoutException) {
+                futureData.cancel(true);
+                PluginManager.CLIENT_REGISTRY.clearLootData();
+                LOGGER.error("Failed to received data: Inactivity timeout occurred. Registration aborted!");
+            } else {
+                LOGGER.error("Failed to finish registering data with error {}", cause.getMessage());
+                cause.printStackTrace();
+            }
         } catch (Throwable e) {
             e.printStackTrace();
-            LOGGER.error("Failed to finish registering data with error {}", e.getMessage());
+            LOGGER.error("Failed to finish registering data with unexpected error {}", e.getMessage());
         }
     }
 
@@ -149,15 +134,9 @@ public class JeiCompatibility implements IModPlugin {
                             }
                         }
 
-                        if (recipeType == null) {
-                            if (LootCategories.BLOCK_LOOT.isHidden()) {
-                                return;
-                            }
-
-                            recipeType = defaultBlockLoot.getValue().getRecipeType();
-                        }
-
+                        if (recipeType != null) {
                         blockRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new BlockLootType(block, node, Collections.emptyList(), outputs));
+                        }
                     },
                     (node, location, entity, outputs) -> {
                         IRecipeType<RecipeHolder<EntityLootType>> recipeType = null;
@@ -173,21 +152,15 @@ public class JeiCompatibility implements IModPlugin {
                             }
                         }
 
-                        if (recipeType == null) {
-                            if (LootCategories.ENTITY_LOOT.isHidden()) {
-                                return;
-                            }
-
-                            recipeType = defaultEntityLoot.getValue().getRecipeType();
+                        if (recipeType != null) {
+                            entityRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new EntityLootType(entity, location, node, Collections.emptyList(), outputs));
                         }
-
-                        entityRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new EntityLootType(entity, location, node, Collections.emptyList(), outputs));
                     },
                     (node, location, outputs) -> {
                         IRecipeType<RecipeHolder<GameplayLootType>> recipeType = null;
 
                         for (JeiGameplayLoot recipeCategory : gameplayCategories.values()) {
-                            if (recipeCategory.getLootCategory().validate(location.getPath())) {
+                            if (recipeCategory.getLootCategory().validate(location)) {
                                 if (recipeCategory.getLootCategory().isHidden()) {
                                     return;
                                 }
@@ -197,21 +170,15 @@ public class JeiCompatibility implements IModPlugin {
                             }
                         }
 
-                        if (recipeType == null) {
-                            if (LootCategories.GAMEPLAY_LOOT.isHidden()) {
-                                return;
-                            }
-
-                            recipeType = defaultGameplayLoot.getValue().getRecipeType();
+                        if (recipeType != null) {
+                            gameplayRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new GameplayLootType(node, location.getPath(), Collections.emptyList(), outputs));
                         }
-
-                        gameplayRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new GameplayLootType(node, location.getPath(), Collections.emptyList(), outputs));
                     },
                     (node, location, inputs, outputs) -> {
                         IRecipeType<RecipeHolder<TradeLootType>> recipeType = null;
 
                         for (JeiTradeLoot recipeCategory : tradeCategories.values()) {
-                            if (recipeCategory.getLootCategory().validate(location.getPath())) {
+                            if (recipeCategory.getLootCategory().validate(location)) {
                                 if (recipeCategory.getLootCategory().isHidden()) {
                                     return;
                                 }
@@ -221,21 +188,15 @@ public class JeiCompatibility implements IModPlugin {
                             }
                         }
 
-                        if (recipeType == null) {
-                            if (LootCategories.TRADE_LOOT.isHidden()) {
-                                return;
-                            }
-
-                            recipeType = defaultTradeLoot.getValue().getRecipeType();
+                        if (recipeType != null) {
+                            tradeRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new TradeLootType(node, location.getPath(), inputs, outputs));
                         }
-
-                        tradeRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new TradeLootType(node, location.getPath(), inputs, outputs));
                     },
                     (node, location, inputs, outputs) -> {
                         IRecipeType<RecipeHolder<TradeLootType>> recipeType = null;
 
                         for (JeiTradeLoot recipeCategory : tradeCategories.values()) {
-                            if (recipeCategory.getLootCategory().validate(location.getPath())) {
+                            if (recipeCategory.getLootCategory().validate(location)) {
                                 if (recipeCategory.getLootCategory().isHidden()) {
                                     return;
                                 }
@@ -245,15 +206,9 @@ public class JeiCompatibility implements IModPlugin {
                             }
                         }
 
-                        if (recipeType == null) {
-                            if (LootCategories.TRADE_LOOT.isHidden()) {
-                                return;
-                            }
-
-                            recipeType = defaultTradeLoot.getValue().getRecipeType();
+                        if (recipeType != null) {
+                            tradeRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new TradeLootType(node, location.getPath(), inputs, outputs));
                         }
-
-                        tradeRecipeTypes.computeIfAbsent(recipeType, (p) -> new LinkedList<>()).add(new TradeLootType(node, location.getPath(), inputs, outputs));
                     }
             );
 
@@ -287,34 +242,21 @@ public class JeiCompatibility implements IModPlugin {
         //noinspection unchecked
         IRecipeType<RecipeHolder<V>> recipeType = (IRecipeType<RecipeHolder<V>>) (Object) IRecipeType.create(e.getKey().getNamespace(), e.getKey().getPath(), RecipeHolder.class);
         Component title = Component.translatable("emi.category." + e.getKey().getNamespace() + "." + e.getKey().getPath().replace('/', '.'));
-        return constructor.construct(guiHelper, recipeType, e, title, guiHelper.createDrawableItemStack(e.getIcon()));
+        return constructor.construct(guiHelper, recipeType, e, title, guiHelper.createDrawableItemStack(e.getIcon().getDefaultInstance()));
     }
 
-    private static <T, U, V extends IType> T createCategory(IGuiHelper guiHelper, Map.Entry<ResourceLocation, LootCategory<U>> e, LootConstructor<T, U, V> constructor) {
-        //noinspection unchecked
-        ResourceLocation id = e.getKey();
-        IRecipeType<RecipeHolder<V>> recipeType = (IRecipeType<RecipeHolder<V>>) (Object) IRecipeType.create(id.getNamespace(), id.getPath(), RecipeHolder.class);
-        Component title = Component.translatable("emi.category." + id.getNamespace() + "." + id.getPath().replace('/', '.'));
-        return constructor.construct(guiHelper, recipeType, e.getValue(), title, guiHelper.createDrawableItemStack(e.getValue().getIcon()));
+    @NotNull
+    private static  <T, U, V extends IType> Collector<LootCategory<U>, ?, Map<LootCategory<U>, T>> getCollector(IGuiHelper guiHelper, LootConstructor<T, U, V> supplier) {
+        return Collectors.toMap(
+                (e) -> e,
+                (e) -> createCategory(guiHelper, e, supplier),
+                (a, b) -> a,
+                LinkedHashMap::new
+        );
     }
 
     @FunctionalInterface
     private interface LootConstructor<T, U, V extends IType> {
         T construct(IGuiHelper guiHelper, IRecipeType<RecipeHolder<V>> recipeType, LootCategory<U> lootCategory, Component title, IDrawable icon);
-    }
-
-    @NotNull
-    private <T extends IRecipeCategory<?>, U, V extends IType> T getDefaultCategory(IGuiHelper guiHelper, LootConstructor<T, U, V> constructor,
-                                                                                    IRecipeCategoryRegistration registry, Map<LootCategory<U>, T> categories,
-                                                                                    LootCategory<U> defaultCategory) {
-        T recipeCategory = categories.remove(defaultCategory);
-
-        if (recipeCategory != null) {
-            return recipeCategory;
-        } else {
-            T defaultRecipeCategory = createCategory(guiHelper, defaultCategory, constructor);
-            registry.addRecipeCategories(defaultRecipeCategory);
-            return defaultRecipeCategory;
-        }
     }
 }