+++ fabric/src/main/java/com/yanny/ali/platform/FabricPlatformHelper.java
@@ -7,6 +7,7 @@ import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 import org.slf4j.Logger;
 
+import java.nio.file.Path;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -33,5 +34,10 @@ public class FabricPlatformHelper implements IPlatformHelper {
 
         LOGGER.info("Found {} plugin(s)", plugins.size());
         return plugins;
+    }
+
+    @Override
+    public Path getConfiguration() {
+        return FabricLoader.getInstance().getConfigDir();
     }
 }